+++ services/logme/model_get_metrics_response.go
@@ -25,9 +25,17 @@ type GetMetricsResponse struct {
 	// REQUIRED
 	DiskPersistentUsed *int64 `json:"diskPersistentUsed"`
 	// REQUIRED
+	Load1 *int64 `json:"load1"`
+	// REQUIRED
+	Load15 *int64 `json:"load15"`
+	// REQUIRED
+	Load5 *int64 `json:"load5"`
+	// REQUIRED
 	MemoryTotal *int64 `json:"memoryTotal"`
 	// REQUIRED
 	MemoryUsed *int64 `json:"memoryUsed"`
+	// REQUIRED
+	OpensearchDashboardURL *string `json:"opensearchDashboardURL"`
 	// REQUIRED
 	ParachuteDiskEphemeralActivated *int64 `json:"parachuteDiskEphemeralActivated"`
 	// REQUIRED